remove tracing abstract and create setup jaeger for clean code
